allow data in create_array
